Fix userpage Chart
